chore: Update database queries to include recipe category name
This commit updates the database queries in the `DatabaseHelper` class to include the name of the recipe category. This is achieved by joining the `recipes` table with the `recipe_categories` table and selecting the category name. The updated queries are used in the `queryAllRecipes` and `getRecipe` methods.

It also adds a new method `updateRecipeStarred` to the `DatabaseHelper` class, which allows updating the `is_starred` field of a recipe. This method is used in the `_toggleStarred` method of the `RecipePreviewPage` class to toggle the starred status of a recipe.

These changes improve the functionality and user experience of the app by providing more information about the recipe category and allowing users to star their favorite recipes.